Fixed bug in Plant2 (with help from Rob).  Was throwing "Mass Balance Error in Photosynthesis DM Allocation" when ever there was excess biomass to be allocated to the sink organ.  This was because in AnnualPlantArbitrator.cs the test in lines 116 and 418 .... if (Organs[i].Name == DMSink.ValueString ... was case sensitive and not allocating excess DM if the sink term did not have a capitial.
To fix this we replaced it with... if (string.Compare(Organs[i].Name, DMSink.ValueString, true) == 0) ... which is not case sensitive.   

This may cause diffs in other plant2 tests by fixing errors

It has just dawned on me that this bug will be present in the Nallocation code as well but we haven't fixed that yet.

Made some more changes to potato2.xml.

Added undergroundorgan into the plant2ide toolbox