This is trying to tidy up diffs created by 1417 (message below) which bob did not complete running
Sorry it this dirties the build.  I am trying to get bob to stop rerunning my patch.

Committing modifications to plant2

Changed RUE model in source code and all crop2.xml files to give developers more flexibility in setting up crop RUE responses

Other changes are a step along the path to improving N and DM partitioning.  None of the current tests in the build are using any of these bits.  That is the next job.